+++ src/main/java/com/github/mjkuranda/spaceadventure2/GameData.java
@@ -121,7 +121,7 @@ public class GameData {
         }
 
         /** Player update */
-        PlayerStatistics.getInstance().addDistance(player.getSpeed());
+        PlayerData.getInstance().addDistance(player.getSpeed());
 
         Input in = container.getInput();
 
@@ -143,16 +143,16 @@ public class GameData {
 
         /** Player collision */
         Entity e = playerCollides();
-        PlayerStatistics.getInstance().unvibrate();
+        PlayerData.getInstance().unvibrate();
 
         if (e != null) {
             player.damage(e.getDurability());
             spawn(new Particle(GameAnimation.EXPLOSION, e));
             e.remove();
-            PlayerStatistics.getInstance().vibrate();
+            PlayerData.getInstance().vibrate();
 
             if (!player.isAlive()) {
-                PlayerStatistics stats = PlayerStatistics.getInstance();
+                PlayerData stats = PlayerData.getInstance();
                 HighScoreHandler.getInstance().inputRecord(new HighScoreRecord(stats.getName(), stats.getScore(), LocalDate.now().toString()));
 
                 game.enterState(StatesId.GAME_OVER_MENU);